Revert "created user settings page"
