+++ .env
@@ -1,2 +1,2 @@
-DATABASE=mongodb+srv://admin:123@cluster0-e6tfj.mongodb.net/test?retryWrites=true&w=majority
+DATABASE=mongodb://localhost:27017/websocket-messenger
 SECRET=SUPERSECRETPASSWORD123
