+++ .gitignore
@@ -3,6 +3,7 @@
 /client/node_modules
 
 .DS_Store
+.env
 
 npm_debug.log*
 yarn_debug.log*
