--- client/src/Commons/SubmitFormButton.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react'
-
-export interface ISubmitFormButtonProps {}
-
-export default class SubmitFormButton extends React.Component<
-    ISubmitFormButtonProps
-> {
-    public render() {
-        return <div></div>
-    }
-}
