+++ client/src/components/Auth/SignIn.tsx
@@ -1,19 +1,16 @@
 import * as React from 'react'
 import { inject, observer } from 'mobx-react'
 import { Link } from 'react-router-dom'
-import { IGetStore } from '../../store/MainStore'
+import MainStore, { IRegisterProps } from '../../store/MainStore'
 import { action, observable } from 'mobx'
-import { IRegisterProps } from './store/Auth.store'
 
 export interface ISignInProps {}
-@inject('getStore')
+@inject('store')
 @observer
 export default class SignIn extends React.Component<ISignInProps> {
-    mainStore = this.injected.getStore('mainStore')
-    authStore = this.injected.getStore('authStore')
-
+    mainStore = this.injected.store
     private get injected() {
-        return this.props as ISignInProps & IGetStore
+        return this.props as ISignInProps & { store: MainStore }
     }
 
     @observable
@@ -21,19 +18,19 @@ export default class SignIn extends React.Component<ISignInProps> {
 
     @action.bound
     onChangeSaveValue(e: React.FormEvent<HTMLInputElement>) {
-        const { saveInputValueRegisterForm } = this.authStore
+        const { saveInputValue } = this.mainStore
         const value = e.currentTarget.value
         const prop = e.currentTarget.dataset.id
         if (prop) {
-            saveInputValueRegisterForm(prop as keyof IRegisterProps, value)
+            saveInputValue(prop as keyof IRegisterProps, value)
         }
     }
 
     @action.bound
     onSubmitLogin(e: React.FormEvent<HTMLFormElement>) {
         e.preventDefault()
-        const { clientLogin } = this.authStore
-        const { email, password } = this.authStore.clientRegisterProps
+        const { clientLogin } = this.mainStore
+        const { email, password } = this.mainStore.clientRegisterProps
         if (email && password) {
             this.requiredDataWarning = false
             clientLogin()
@@ -42,36 +39,26 @@ export default class SignIn extends React.Component<ISignInProps> {
         }
     }
     public render() {
-        const { email, password } = this.authStore.clientRegisterProps
+        const { email, password } = this.mainStore.clientRegisterProps
         return (
-            <div className="auth-container form">
+            <div className="auth-container">
                 <p>Welcome to Messenger</p>
                 <form className="auth-form" onSubmit={this.onSubmitLogin}>
-                    <div className="form-row">
-                        <div className="form-label">Email address</div>
-                        <input
-                            type="email"
-                            data-id="email"
-                            value={email}
-                            onChange={this.onChangeSaveValue}
-                        />
-                    </div>
-                    <div className="form-row">
-                        <div className="form-label">Password</div>
-                        <input
-                            type="password"
-                            data-id="password"
-                            value={password}
-                            onChange={this.onChangeSaveValue}
-                        />
-                    </div>
-                    <div className="form-row">
-                        <input
-                            className="button button--primary"
-                            type="submit"
-                            value="Sign In"
-                        />
-                    </div>
+                    <div>Email address</div>
+                    <input
+                        type="email"
+                        data-id="email"
+                        value={email}
+                        onChange={this.onChangeSaveValue}
+                    />
+                    <div>Password</div>
+                    <input
+                        type="password"
+                        data-id="password"
+                        value={password}
+                        onChange={this.onChangeSaveValue}
+                    />
+                    <input type="submit" value="Sign In" />
                 </form>
 
                 {this.requiredDataWarning && (
@@ -81,7 +68,7 @@ export default class SignIn extends React.Component<ISignInProps> {
                 )}
 
                 <div className="create-account">
-                    New to messenger?&nbsp;
+                    New to messenger?
                     <Link to={'/register'}>Create an account</Link>
                 </div>
             </div>
