+++ client/src/components/Mainchatpage/Logout/Logout.tsx
@@ -1,26 +1,26 @@
 import * as React from 'react'
 import { inject, observer } from 'mobx-react'
 import { action } from 'mobx'
-import MainStore, { IGetStore } from '../../../store/MainStore'
+import MainStore from '../../../store/MainStore'
 
 export interface ILogoutProps {}
 
-@inject('getStore')
+@inject('store')
 @observer
 export default class Logout extends React.Component<ILogoutProps> {
-    authStore = this.injected.getStore('authStore')
+    mainStore = this.injected.store
 
     private get injected() {
-        return this.props as ILogoutProps & IGetStore
+        return this.props as ILogoutProps & { store: MainStore }
     }
     @action.bound
     onClickLogout() {
-        const { logout } = this.authStore
+        const { logout } = this.mainStore
         logout()
     }
     public render() {
         return (
-            <div className="logout-container">
+            <div className="logout__container">
                 <div onClick={this.onClickLogout}>Logout</div>
             </div>
         )
