+++ client/src/requests/index.ts
@@ -1,5 +1,4 @@
-import { IChangeSettingsProps } from './../components/Mainchatpage/SettingsBar/store/Settings.store'
-import { IRegisterProps } from './../components/Auth/store/Auth.store'
+import { IRegisterProps } from './../store/MainStore'
 
 export interface IRegisterResponse {
     success: string
@@ -12,16 +11,10 @@ export interface IAuthResponse {
     success: string
     isOnline: boolean
     id: string
-    email: string
     name: string
     lastname: string
     avatarUrl: string
     dialogs: []
-}
-
-export interface ISaveProfilePhotoResponse {
-    avatarUrl: any
-    _id: string
 }
 
 export function registerClient(clientData: IRegisterProps) {
@@ -69,46 +62,6 @@ export function authClientReq(token: IAuthResponse) {
         return fetch('/api/users/auth', {
             method: 'POST',
             body: JSON.stringify({ token }),
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        }).then((res) => {
-            res.json().then((parsedRes) => {
-                if (res.status === 200) {
-                    resolve(parsedRes)
-                } else {
-                    reject(parsedRes)
-                }
-            })
-        })
-    })
-}
-
-export function saveProfilePhotoReq(data: ISaveProfilePhotoResponse) {
-    return new Promise<ISaveProfilePhotoResponse>((resolve, reject) => {
-        return fetch('/api/users/addAvatar', {
-            method: 'POST',
-            body: JSON.stringify(data),
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        }).then((res) => {
-            res.json().then((parsedRes) => {
-                if (res.status === 200) {
-                    resolve(parsedRes)
-                } else {
-                    reject(parsedRes)
-                }
-            })
-        })
-    })
-}
-
-export function updateClientSettings(data: IChangeSettingsProps) {
-    return new Promise<IChangeSettingsProps>((resolve, reject) => {
-        return fetch('/api/users/updateClient', {
-            method: 'POST',
-            body: JSON.stringify(data),
             headers: {
                 'Content-Type': 'application/json',
             },
