+++ client/src/style.scss
@@ -1,7 +1,3 @@
-$main-color: #4d9670;
-$main-color-blue: #4d7d96;
-$main-color-light-blue: #71a2bb;
-
 html,
 body,
 div,
@@ -18,61 +14,6 @@ body,
     height: 100%;
 }
 
-input[type='text'],
-input[type='email'],
-input[type='password'] {
-    border: 1px solid #c1c1c1;
-    padding: 12px 8px;
-    border-radius: 5px;
-    width: 100%;
-    box-sizing: border-box;
-
-    &::placeholder {
-        font-style: italic;
-        color: #a9a9a9;
-    }
-}
-
-.font-small {
-    font-size: 0.8em;
-}
-
-.button {
-    border: 1px solid #a7a7a7;
-    background: #f3f3f3;
-    padding: 10px 20px;
-    border-radius: 5px;
-    text-decoration: none;
-    color: inherit;
-    cursor: pointer;
-
-    &--primary {
-        background: $main-color-blue;
-        border-color: darken($color: $main-color-blue, $amount: 10);
-        color: white;
-    }
-
-    &--small {
-        font-size: 0.8em;
-        padding: 7px 15px;
-    }
-}
-
-.form {
-    p {
-        font-size: 1.8em;
-    }
-}
-
-.form-row {
-    margin-bottom: 15px;
-}
-
-.form-label {
-    color: gray;
-    margin-bottom: 5px;
-}
-
 .wrapper {
     height: 100%;
     width: 100%;
@@ -85,7 +26,7 @@ input[type='password'] {
 .header {
     width: 100%;
     height: 160px;
-    background-color: $main-color-blue;
+    background-color: rgb(77, 150, 112);
 }
 
 .main {
@@ -97,7 +38,6 @@ input[type='password'] {
 }
 
 .app-container {
-    position: relative;
     width: 1396px;
     height: 95%;
     background-color: white;
@@ -122,118 +62,7 @@ input[type='password'] {
     &__header {
         @extend .chat-header;
         border-right: 1px solid #c5c5c5;
-        display: flex;
     }
-
-    &__header-settings-button {
-        width: 50px;
-        height: 50px;
-        padding: 10px;
-        cursor: pointer;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        transition: all 0.3s ease;
-
-        &:hover {
-            background: rgba(0, 0, 0, 0.08);
-        }
-    }
-}
-
-.client-settings {
-    width: 100%;
-    border-right: 1px solid #ededed;
-    display: flex;
-    flex-direction: column;
-
-    &__header {
-        display: flex;
-        background-color: $main-color-light-blue;
-        height: 70px;
-        color: white;
-    }
-
-    &__header-title {
-        font-size: 23px;
-        line-height: 75px;
-        padding-left: 30px;
-    }
-
-    &__header-arrow {
-        padding: 25px;
-        cursor: pointer;
-        transition: all 0.3s ease;
-
-        &:hover {
-            background: rgba(0, 0, 0, 0.08);
-        }
-    }
-
-    &__client-pic-container {
-        width: 100%;
-        height: 230px;
-    }
-
-    &__client-pic {
-        background-color: lightgray;
-        margin: 25px auto;
-        background-position: center center;
-        background-repeat: no-repeat;
-        background-size: contain;
-        border-radius: 50%;
-        cursor: pointer;
-        width: 200px;
-        height: 200px;
-        position: relative;
-
-        &:hover .client-settings__client-pic-shadow-container {
-            display: block;
-        }
-    }
-    &__client-pic-shadow-container {
-        width: 100%;
-        height: 100%;
-        display: none;
-        background-color: #9e9e9eab;
-        border-radius: 50%;
-        cursor: pointer;
-        position: absolute;
-        z-index: 5;
-    }
-    &__upload-input {
-        opacity: 0;
-    }
-    &__upload-label {
-        font-size: 17px;
-        text-align: center;
-        display: none;
-        width: 100%;
-        height: 100%;
-        display: block;
-        border-radius: 50%;
-    }
-    &__data-change-form {
-        margin: 30px auto;
-        width: 300px;
-    }
-    &__data-change-box {
-        margin: 10px 0;
-    }
-    &__data-change-title {
-        font-size: 17px;
-    }
-    &__data-change-prop {
-        font-size: 16px;
-        display: inline-block;
-    }
-}
-
-.logout-container {
-    font-size: 23px;
-    line-height: 75px;
-    margin-left: 30px;
-    cursor: pointer;
 }
 
 .dialog {
@@ -270,7 +99,7 @@ input[type='password'] {
         border: none;
         border-radius: 10px;
         border: 1px solid #cccccc;
-        padding: 3px 80px 3px 55px !important;
+        padding: 3px 80px 3px 55px;
         font-size: 15px;
 
         &:focus {
@@ -287,7 +116,8 @@ input[type='password'] {
         width: 6%;
         height: 50%;
         margin: 15px 20px;
-        background: $main-color-blue;
+        border: 1px solid rgb(77, 150, 112);
+        background: #4d9670;
         text-align: center;
         position: absolute;
         width: 50px;
@@ -377,8 +207,8 @@ input[type='password'] {
 }
 
 .create-account {
-    margin: 40px auto;
-    max-width: 400px;
+    margin: 0 auto;
+    max-width: 600px;
     &__title {
         margin-bottom: 30px;
         text-align: center;
@@ -386,13 +216,10 @@ input[type='password'] {
     &__form {
         display: flex;
         flex-direction: column;
+        max-width: 500px;
     }
 
     &__warning {
         color: #e21515;
     }
-}
-
-.auth-container {
-    margin: 40px auto;
 }
