+++ package.json
@@ -22,7 +22,6 @@
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^3.5.5",
-    "mongoose": "^5.9.7",
-    "nodemon": "^2.0.3"
+    "mongoose": "^5.9.7"
   }
 }
